put solvers in one file
